fix(auth): check token validity before calling the API

+++ lib-homeassistant-api/src/main/java/fr/outadoc/homeslide/hassapi/model/auth/Token.kt
@@ -8,7 +8,7 @@ data class Token(
     @Json(name = "access_token")
     val accessToken: String,
     @Json(name = "expires_in")
-    val expiresIn: Int,
+    val expiresIn: Long,
     @Json(name = "token_type")
     val tokenType: String,
     @Json(name = "refresh_token")
